refactor: 이전 리뷰 사항(패키지명 변경, 애너테이션 순서 변경) 반영
